Replace swift.py with remote_swift.py
